Split the engine proper out into a framework

+++ UAGithubEngine.m
@@ -661,7 +661,7 @@
     
     // Get response code.
     NSHTTPURLResponse *resp = (NSHTTPURLResponse *)response;
-    int statusCode = resp.statusCode;
+    NSInteger statusCode = resp.statusCode;
 	
 	
 	if ([[[resp allHeaderFields] allKeys] containsObject:@"X-Ratelimit-Remaining"] && [[[resp allHeaderFields] valueForKey:@"X-Ratelimit-Remaining"] isEqualToString:@"1"])
